refactor(tab-panels): 폴더 생성 후 index.tsx로 파일명 변경
